change the FPS count + refactor
move calculation of FPS to the begin of animate() and
now it checks if the 1000ms is passed instead of 60 frames.
Organize some properties of currentStatus into objects.
Change perfomance.now to event.timeStamp in start and restart listeners.
Move some parts of the code to functions (gameOver(), resetHearts(),
togglePauseButton(pauseBtn, pauseBtnText, text, classToRemove, classToAdd))